scientific_parser/agents/requirements.txt
pydantic-ai 
mcp 
a2a-python
pydantic-graph
arxiv
google-generativeai
python-dotenv
requests
rich
asyncclick